pyproject.toml
[project]
name = "fx-thing-fxprocessor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "jack-client>=0.5.5",
    "pedalboard>=0.9.22",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pip>=26.0.1",
    "ruff>=0.15.8",
]
